requirements.txt
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==24.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.2
